deleted terms test in cse_test and added new file terms_test
